refactor: simplify premium integrations handling
